Move ExecutionPolicy and Plugin into org.renaissance package
These are potentially of interest to harness users, and we don't want
them in the "harness" package loaded by a different class loader.
