+++ renaissance-harness/src/main/java/org/renaissance/ExecutionDriver.java
@@ -1,7 +1,6 @@
 package org.renaissance;
 
 import org.renaissance.BenchmarkResult.ValidationException;
-import org.renaissance.harness.ExecutionPolicy;
 import org.renaissance.core.BenchmarkInfo;
 
 /**
