+++ renaissance-harness/src/main/scala/org/renaissance/MarkdownGenerator.scala
@@ -9,9 +9,7 @@ import org.renaissance.core.BenchmarkInfo
 import org.renaissance.core.BenchmarkRegistry
 import org.renaissance.core.Launcher
 import org.renaissance.core.ModuleLoader
-import org.renaissance.harness.ExecutionPolicy
-import org.renaissance.harness.Plugin
-import org.renaissance.harness.Plugin._
+import org.renaissance.Plugin._
 import scopt.OptionParser
 
 import scala.collection.JavaConverters._
